refactor(cli): extract getAnalysisAge and estimateCost to command-helpers
Both functions were duplicated across CLI commands with diverged
implementations. getAnalysisAge was identical; estimateCost had a
simpler (less accurate) version in run.ts vs the per-stage breakdown
in generate.ts — the accurate version is now the single source of truth.
